chore(ai): the embed lane takes its own file under the length cap
